Version finale de CreateTutor, UpdateTutor et DeleteTutor

+++ FlOpEDT/api_graphql/tutors/delete.py
@@ -1,4 +1,3 @@
-from graphene_django import DjangoObjectType
 import graphene
 from people.models import Tutor
 from .types import TutorType
@@ -6,13 +5,17 @@ from graphql_relay import from_global_id
 
 class DeleteTutor(graphene.Mutation):
         class Arguments:
-            id = graphene.ID()
+            id = graphene.ID(required=True)
 
         tutor = graphene.Field(TutorType)
 
         @classmethod
         def mutate(cls, root, info, id):
-            _, object_id = from_global_id(id)
-            tutor = Tutor.objects.get(id=object_id)
-            tutor.delete()
-            return DeleteTutor(tutor)
+            id = from_global_id(id) [1]
+            tutor_set = Tutor.objects.filter(id=id)
+            if tutor_set:
+                tutor=tutor_set.first()
+                tutor.delete()
+                return DeleteTutor(tutor)
+            else:
+                print('Tutor with given ID does not exist in the database')
